Move exception checks to colmap/util/logging.h
