+++ src/colmap/util/logging.h
@@ -31,7 +31,9 @@
 
 #include "colmap/util/string.h"
 
+#include <exception>
 #include <iostream>
+#include <string>
 
 #include <glog/logging.h>
 
@@ -60,6 +62,42 @@
 #define CHECK_OPTION_LT(val1, val2) CHECK_OPTION_OP(_LT, <, val1, val2)
 #define CHECK_OPTION_GE(val1, val2) CHECK_OPTION_OP(_GE, >=, val1, val2)
 #define CHECK_OPTION_GT(val1, val2) CHECK_OPTION_OP(_GT, >, val1, val2)
+
+// Option checker macros. In contrast to glog, this function does not abort the
+// program, but simply throws an exception on failure.
+#define THROW_EXCEPTION(exception, msg) \
+  throw TemplateException<exception>(__FILE__, __LINE__, ToString(msg));
+
+#define THROW_CUSTOM_CHECK_MSG(condition, exception, msg) \
+  if (!(condition))                                       \
+    throw TemplateException<exception>(                   \
+        __FILE__,                                         \
+        __LINE__,                                         \
+        __GetCheckString(#condition) + std::string(" ") + ToString(msg));
+
+#define THROW_CUSTOM_CHECK(condition, exception) \
+  if (!(condition))                              \
+    throw TemplateException<exception>(          \
+        __FILE__, __LINE__, __GetCheckString(#condition));
+
+#define THROW_CHECK(expr) __ThrowCheckImpl(__FILE__, __LINE__, !!(expr), #expr);
+
+#define THROW_CHECK_MSG(expr, msg) \
+  __ThrowCheckImplMsg(__FILE__, __LINE__, !!(expr), #expr, ToString(msg))
+
+#define THROW_CHECK_NOTNULL(val) \
+  __ThrowCheckNotNull(__FILE__, __LINE__, (val), #val)
+
+#define THROW_CHECK_OP(name, op, val1, val2) \
+  __ThrowCheckOpImpl(                        \
+      __FILE__, __LINE__, (val1 op val2), val1, val2, #val1, #val2, #op);
+
+#define THROW_CHECK_EQ(val1, val2) THROW_CHECK_OP(_EQ, ==, val1, val2)
+#define THROW_CHECK_NE(val1, val2) THROW_CHECK_OP(_NE, !=, val1, val2)
+#define THROW_CHECK_LE(val1, val2) THROW_CHECK_OP(_LE, <=, val1, val2)
+#define THROW_CHECK_LT(val1, val2) THROW_CHECK_OP(_LT, <, val1, val2)
+#define THROW_CHECK_GE(val1, val2) THROW_CHECK_OP(_GE, >=, val1, val2)
+#define THROW_CHECK_GT(val1, val2) THROW_CHECK_OP(_GT, >, val1, val2)
 
 namespace colmap {
 
@@ -98,6 +136,88 @@ bool __CheckOptionOpImpl(const char* file,
                                std::to_string(val1).c_str(),
                                std::to_string(val2).c_str());
     return false;
+  }
+}
+
+template <typename T>
+inline std::string ToString(T msg) {
+  return std::to_string(msg);
+}
+
+inline std::string ToString(std::string msg) { return msg; }
+
+inline std::string ToString(const char* msg) { return std::string(msg); }
+
+template <typename T>
+inline T TemplateException(const char* file,
+                           const int line,
+                           const std::string& txt) {
+  std::stringstream ss;
+  ss << "[" << __GetConstFileBaseName(file) << ":" << line << "] " << txt;
+  // TODO: cleanup
+  return T(ss.str());
+}
+
+inline std::string __GetCheckString(const char* cond_str) {
+  // TODO: cleanup
+  std::stringstream ss;
+  ss << "Check Failed: " << cond_str;
+  return ss.str();
+}
+
+inline void __ThrowCheckImpl(const char* file,
+                             const int line,
+                             const bool result,
+                             const char* expr_str) {
+  if (!result) {
+    throw TemplateException<std::invalid_argument>(
+        file, line, __GetCheckString(expr_str).c_str());
+  }
+}
+
+inline void __ThrowCheckImplMsg(const char* file,
+                                const int line,
+                                const bool result,
+                                const char* expr_str,
+                                const std::string& msg) {
+  if (!result) {
+    std::string m = std::string(expr_str) + " : " + msg;
+    throw TemplateException<std::invalid_argument>(
+        file, line, __GetCheckString(m.c_str()));
+  }
+}
+
+template <typename T>
+T __ThrowCheckNotNull(const char* file,
+                      const int line,
+                      T&& t,
+                      const char* name) {
+  if (t == nullptr) {
+    std::string msg = "\"" + std::string(name) + "\" Must be non nullptr";
+    throw TemplateException<std::invalid_argument>(
+        file, line, __GetCheckString(msg.c_str()));
+  } else {
+    return std::forward<T>(t);
+  }
+}
+
+template <typename T1, typename T2>
+void __ThrowCheckOpImpl(const char* file,
+                        const int line,
+                        const bool result,
+                        const T1& val1,
+                        const T2& val2,
+                        const char* val1_str,
+                        const char* val2_str,
+                        const char* op_str) {
+  if (!result) {
+    // TODO: cleanup
+    std::stringstream ss;
+    ss << val1_str << " " << op_str << " " << val2_str << " (" << val1
+       << " vs. " << val2 << ")";
+    std::string msg = ss.str();
+    throw TemplateException<std::invalid_argument>(
+        file, line, __GetCheckString(msg.c_str()));
   }
 }
