+++ src/pycolmap/estimators/alignment.h
@@ -4,8 +4,7 @@
 #include "colmap/exe/model.h"
 #include "colmap/geometry/sim3.h"
 #include "colmap/scene/reconstruction.h"
-
-#include "pycolmap/log_exceptions.h"
+#include "colmap/util/logging.h"
 
 #include <pybind11/eigen.h>
 #include <pybind11/pybind11.h>
