+++ src/pycolmap/estimators/fundamental_matrix.h
@@ -4,8 +4,8 @@
 #include "colmap/math/random.h"
 #include "colmap/optim/loransac.h"
 #include "colmap/scene/camera.h"
+#include "colmap/util/logging.h"
 
-#include "pycolmap/log_exceptions.h"
 #include "pycolmap/utils.h"
 
 #include <pybind11/eigen.h>
