+++ src/pycolmap/estimators/generalized_absolute_pose.h
@@ -6,8 +6,8 @@
 #include "colmap/math/random.h"
 #include "colmap/optim/ransac.h"
 #include "colmap/scene/camera.h"
+#include "colmap/util/logging.h"
 
-#include "pycolmap/log_exceptions.h"
 #include "pycolmap/utils.h"
 
 #include <pybind11/eigen.h>
