+++ src/pycolmap/estimators/triangulation.h
@@ -3,9 +3,9 @@
 #include "colmap/estimators/triangulation.h"
 #include "colmap/scene/camera.h"
 #include "colmap/scene/image.h"
+#include "colmap/util/logging.h"
 
 #include "pycolmap/helpers.h"
-#include "pycolmap/log_exceptions.h"
 #include "pycolmap/utils.h"
 
 #include <pybind11/eigen.h>
