+++ src/pycolmap/pipeline/sfm.h
@@ -8,7 +8,6 @@
 #include "colmap/util/misc.h"
 
 #include "pycolmap/helpers.h"
-#include "pycolmap/log_exceptions.h"
 
 #include <memory>
