+++ src/pycolmap/scene/point2D.h
@@ -1,11 +1,11 @@
 #pragma once
 
 #include "colmap/scene/point2d.h"
+#include "colmap/util/logging.h"
 #include "colmap/util/misc.h"
 #include "colmap/util/types.h"
 
 #include "pycolmap/helpers.h"
-#include "pycolmap/log_exceptions.h"
 
 #include <memory>
 #include <sstream>
